Added a filter system to brushes, and moved color calculation to an Ink trait.
Created a filter that calculates the angle for the stroke points based on their direction.

Also created a filter that calculates undulating color based on the time, and changed the default ink to use the colors returned by it.

Some cracks are visible between segments, and as expected there's problems with calculating the angle when sample points are close to each other.