Added a ZeroLengthSegmentFilter, which discards any stroke points that did not move from the previous position and merges their other properties with the next point.  This fixed the annoying issue with blobs (issue #3).
Cleaned up the StrokeAngleTilter class a little bit, and made the smoothing property public.  With it set close to 1 (around 0.8 or so), it gives a nice calligraphy style tilt to the pen without locking it in a fixed awkward angle like most calligraphy brushes tend to do.  Set it to no smoothing by default though.