Fixed triangle based stroke renderer.  It's fast, but there's warping problems e.g. when stroke size is changing.  It is also not handling subpixel accuracy as nicely as the previous renderer, but on the other hand it gets rid of many artefacts.
Did various optimizations to inner loops and often called functions, such as the data sample interpolation and color alpha mixing.

Changed accesses of comon properties to use pre-calculated integer id:s instead of recalculating the ID:s on each access.  This sped things up especially for the inks that regularily get and set r,g,b,a and many other properties for each pixel.

Rendering is now significantly faster, but there is still some brush overlapping to deal with using a stroke filter, and the triangle renderer could be switched to one that does correct texture mapping and subpixel accuracy, and can do antialiasing along the outer (but not inner) edges.